update posts et profile updated

+++ controllers/post.ctrl.js
@@ -48,20 +48,33 @@ exports.update = async (req, res, next) => {
     console.log(req.body.post);
     console.log(req.body);
     const reqPost = JSON.parse(req.body.post);
-    if (reqPost.poster_id === reqPost.user_id || reqPost.admin === true) {
-        const updatedPost = await Post.findOne({
-            where: { id: req.params.id },
-        });
-        const filename = updatedPost.image.split("/images/")[1];
-        fs.unlink(`images/${filename}`, async () => {
+    if (req.file) {
+        if (reqPost.poster_id === reqPost.user_id || reqPost.admin === true) {
+            const updatedPost = await Post.findOne({
+                where: { id: req.params.id },
+            });
+            const filename = updatedPost.image.split("/images/")[1];
+            fs.unlink(`images/${filename}`, async () => {
+                updatedPost.text = reqPost.text;
+                (updatedPost.image = `${req.protocol}://${req.get(
+                    "host"
+                )}/images/${req.file.filename}`),
+                    await updatedPost
+                        .save()
+                        .then((data) => res.status(200).json({ data }))
+                        .catch((error) => res.status(400).json({ error }));
+            });
+        }
+    } else {
+        if (reqPost.poster_id === reqPost.user_id || reqPost.admin === true) {
+            const updatedPost = await Post.findOne({
+                where: { id: req.params.id },
+            });
             updatedPost.text = reqPost.text;
-            (updatedPost.image = `${req.protocol}://${req.get("host")}/images/${
-                req.file.filename
-            }`),
-                await updatedPost
-                    .save()
-                    .then((data) => res.status(200).json({ data }))
-                    .catch((error) => res.status(400).json({ error }));
-        });
+            await updatedPost
+                .save()
+                .then((data) => res.status(200).json({ data }))
+                .catch((error) => res.status(400).json({ error }));
+        }
     }
 };
