+++ controllers/user.ctrl.js
@@ -98,25 +98,33 @@ exports.updateOne = async (req, res, next) => {
     const updatedUser = JSON.parse(req.body.user);
     const id = updatedUser.user_id;
     const username = updatedUser.username;
-
-    const user2Update = await User.findOne({ where: { id: id } });
-    const filename = user2Update.avatar.split("/images/")[1];
-    if (filename !== "avatar/sun.png") {
-        fs.unlink(`images/${filename}`, async () => {
+    if (req.file) {
+        const user2Update = await User.findOne({ where: { id: id } });
+        const filename = user2Update.avatar.split("/images/")[1];
+        if (filename !== "avatar/sun.png") {
+            fs.unlink(`images/${filename}`, async () => {
+                user2Update.username = username;
+                (user2Update.avatar = `${req.protocol}://${req.get(
+                    "host"
+                )}/images/${req.file.filename}`),
+                    await user2Update
+                        .save()
+                        .then((data) => res.status(200).json({ data }))
+                        .catch((error) => res.status(400).json({ error }));
+            });
+        } else {
             user2Update.username = username;
-            (user2Update.avatar = `${req.protocol}://${req.get(
-                "host"
-            )}/images/${req.file.filename}`),
-                await user2Update
-                    .save()
-                    .then((data) => res.status(200).json({ data }))
-                    .catch((error) => res.status(400).json({ error }));
-        });
+            user2Update.avatar = `${req.protocol}://${req.get("host")}/images/${
+                req.file.filename
+            }`;
+            await user2Update
+                .save()
+                .then((data) => res.status(200).json({ data }))
+                .catch((error) => res.status(400).json({ error }));
+        }
     } else {
+        const user2Update = await User.findOne({ where: { id: id } });
         user2Update.username = username;
-        user2Update.avatar = `${req.protocol}://${req.get("host")}/images/${
-            req.file.filename
-        }`;
         await user2Update
             .save()
             .then((data) => res.status(200).json({ data }))
